refactor(web-ui): finish task detail slice split
